requirements.txt
python-telegram-bot==20.3
google-generativeai
requests
beautifulsoup4
python-dotenv